+++ Mods ChoGGi/Colonist Next Age/Code/Script.lua
@@ -78,4 +78,3 @@ end
 OnMsg.ModsReloaded = ModOptions
 -- Fired when Mod Options>Apply button is clicked
 OnMsg.ApplyModOptions = ModOptions
-